# Back/docs/api_doc/swagger.json
{"openapi":"3.0.1","info":{"title":"'@project.name@'","description":"'@project.description@'","contact":{"name":"John Doe","email":"dumy@dumy"},"version":"'@project.version@'"},"servers":[{"url":"http://localhost:8080","description":"'@project.description@'"}],"paths":{"/api/partidas/{id}/empezar":{"put":{"tags":["partidas-controller"],"operationId":"empezarPartida","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/partidas/{id}/configurar":{"put":{"tags":["partidas-controller"],"operationId":"configPartida","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PartidaConfigDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/partidas/{idPartida}/jugadores":{"put":{"tags":["partidas-controller"],"operationId":"editarJugador","parameters":[{"name":"idPartida","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JugadorDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PartidaModel"}}}}}},"post":{"tags":["partidas-controller"],"operationId":"agregarJugador","parameters":[{"name":"idPartida","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JugadorDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PartidaModel"}}}}}}},"/api/register":{"post":{"tags":["autenticacion-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistroRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/partidas":{"post":{"tags":["partidas-controller"],"operationId":"crearPartida","requestBody":{"content":{"application/json":{"schema":{"type":"integer","format":"int64"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PartidaModel"}}}}}}},"/api/partidas/turnos-refuerzos/refuerzos":{"post":{"tags":["partidas-controller"],"operationId":"recibirRefuerzos","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RefuerzoPostDTO"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/partidas/turnos-ataque/acciones/reagrupaciones":{"post":{"tags":["partidas-controller"],"operationId":"recibirReagrupacion","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccionReagrupacionPostDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccionModel"}}}}}}},"/api/partidas/turnos-ataque/acciones/combates":{"post":{"tags":["partidas-controller"],"operationId":"recibirCombate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccionCombatePostDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccionModel"}}}}}}},"/api/mensajes":{"get":{"tags":["mensajes-controller"],"operationId":"getMensajes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MensajeModel"}}}}}}},"post":{"tags":["mensajes-controller"],"operationId":"addMensaje","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MensajeDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/login":{"post":{"tags":["autenticacion-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioModel"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/tarjetas":{"get":{"tags":["tarjetas-controller"],"operationId":"getTarjetas","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TarjetaModel"}}}}}}}},"/api/partidas/{id}":{"get":{"tags":["partidas-controller"],"operationId":"buscarPartida","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PartidaModel"}}}}}}},"/api/partidas/{id}/rondas/ultima/{idJugador}":{"get":{"tags":["partidas-controller"],"operationId":"buscarUltimaRonda","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idJugador","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RondaModel"}}}}}}},"/api/partidas/{id}/paises":{"get":{"tags":["partidas-controller"],"operationId":"buscarPaises","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaisJugadorModel"}}}}}}}},"/api/partidas/{id}/colores":{"get":{"tags":["partidas-controller"],"operationId":"buscarColores","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ColorModel"}}}}}}}},"/api/partidas/{idPartida}/rondas/{idRonda}/turnos-refuerzos/{idTurno}/finalizar":{"get":{"tags":["partidas-controller"],"operationId":"terminarTurnoRefuerzo","parameters":[{"name":"idPartida","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idRonda","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idTurno","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/partidas/{idPartida}/rondas/{idRonda}/turnos-ataques/{idTurno}/finalizar":{"get":{"tags":["partidas-controller"],"operationId":"terminarTurnoAtaque","parameters":[{"name":"idPartida","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idRonda","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idTurno","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/partidas/usuario/{idUsuario}":{"get":{"tags":["partidas-controller"],"operationId":"obtenerPartidasPorUsuario","parameters":[{"name":"idUsuario","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PartidaModel"}}}}}}}},"/api/partidas/turnos-refuerzos/{id}":{"get":{"tags":["partidas-controller"],"operationId":"obtenerTurnoRefuerzo","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TurnoRefuerzoModel"}}}}}}},"/api/partidas/turnos-refuerzos/{id}/fichas":{"get":{"tags":["partidas-controller"],"operationId":"obtenerFichas","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FichasDTO"}}}}}}},"/api/partidas/turnos-ataques/{id}":{"get":{"tags":["partidas-controller"],"operationId":"obtenerTurnoAtaque","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TurnoAtaqueModel"}}}}}}},"/api/partidas/publicas":{"get":{"tags":["partidas-controller"],"operationId":"obtenerPartidasPublicas","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PartidaModel"}}}}}}}},"/api/partidas/ganador/{id}":{"get":{"tags":["partidas-controller"],"operationId":"chekearGanador","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GanadorDTO"}}}}}}},"/api/paises":{"get":{"tags":["paises-controller"],"operationId":"getAllPaises","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaisModel"}}}}}}}},"/api/paisesJugadores":{"get":{"tags":["paises-jugadores-controller"],"operationId":"getPaisJugador","parameters":[{"name":"idPartida","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"idPais","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaisJugadorDto"}}}}}}},"/api/objetivos":{"get":{"tags":["objetivos-controller"],"operationId":"getObjetivos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ObjetivoModel"}}}}}}}},"/api/objetivos/{id}":{"get":{"tags":["objetivos-controller"],"operationId":"getObjetivo","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ObjetivoModel"}}}}}}},"/api/estadisticas/porcentaje/{idUsuario}":{"get":{"tags":["estadisticas-controller"],"operationId":"porcentajeVictorias","parameters":[{"name":"idUsuario","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"number","format":"double"}}}}}}},"/api/estadisticas/partidas/{idUsuario}":{"get":{"tags":["estadisticas-controller"],"operationId":"partidasPorUsuario","parameters":[{"name":"idUsuario","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/estadisticas/objetivos":{"get":{"tags":["estadisticas-controller"],"operationId":"objetivosCumplidos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}}}}}}}},"/api/estadisticas/colores":{"get":{"tags":["estadisticas-controller"],"operationId":"coloresMasUsados","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}}}}}}}}},"components":{"schemas":{"PartidaConfigDTO":{"type":"object","properties":{"objetivo":{"type":"integer","format":"int32"},"tipo":{"type":"string"}}},"JugadorDTO":{"type":"object","properties":{"idUsuario":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"idColor":{"type":"integer","format":"int32"},"tipoJugador":{"type":"string"}}},"ColorModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"color":{"type":"string"}}},"ContinenteModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"premio":{"type":"integer","format":"int32"}}},"JugadorModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"idPartida":{"type":"integer","format":"int32"},"tipoJugador":{"type":"string","enum":["BOT_NOVATO","BOT_BALANCEADO","BOT_EXPERTO","HUMANO"]},"nombre":{"type":"string"},"color":{"$ref":"#/components/schemas/ColorModel"},"objetivo":{"$ref":"#/components/schemas/ObjetivoModel"},"estadoJugador":{"type":"string","enum":["VIVO","MUERTO"]},"idUsuario":{"type":"integer","format":"int32"},"idJugadorAsesino":{"type":"integer","format":"int32"},"tarjetas":{"type":"array","items":{"$ref":"#/components/schemas/TarjetaModel"}},"paises":{"type":"array","items":{"$ref":"#/components/schemas/PaisJugadorModel"}}}},"ObjetivoCantidadPaisesModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"idObjetivo":{"type":"integer","format":"int32"},"continenteModel":{"$ref":"#/components/schemas/ContinenteModel"},"cantidad":{"type":"integer","format":"int32"}}},"ObjetivoContModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"idObjetivo":{"type":"integer","format":"int32"},"continenteModel":{"$ref":"#/components/schemas/ContinenteModel"}}},"ObjetivoModel":{"type":"object","properties":{"idObjetivo":{"type":"integer","format":"int32"},"colorModel":{"$ref":"#/components/schemas/ColorModel"},"objetivoContModels":{"type":"array","items":{"$ref":"#/components/schemas/ObjetivoContModel"}},"objetivoCantidadPaisesModels":{"type":"array","items":{"$ref":"#/components/schemas/ObjetivoCantidadPaisesModel"}},"id":{"type":"integer","format":"int64"}}},"PaisJugadorModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"idPartida":{"type":"integer","format":"int32"},"fichas":{"type":"integer","format":"int32"},"idJugador":{"type":"integer","format":"int32"},"idPais":{"type":"integer","format":"int32"}}},"PartidaModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"estado":{"type":"string","enum":["EN_PREPARACION","SIN_JUGAR","EN_REFUERZO","EN_ATAQUE","EN_CURSO","FINALIZADO"]},"idUsuario":{"type":"integer","format":"int32"},"cantidadParaGanar":{"type":"integer","format":"int32"},"tipoPartida":{"type":"string","enum":["PUBLICA","PRIVADA"]},"jugadores":{"type":"array","items":{"$ref":"#/components/schemas/JugadorModel"}}}},"TarjetaModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"simbolo":{"type":"string","enum":["BARCO","COMODIN","CANION","GLOBO"]},"estado":{"type":"string"}}},"RegistroRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"repeatPassword":{"type":"string"}}},"RefuerzoPostDTO":{"type":"object","properties":{"idTurnoRef":{"type":"integer","format":"int32"},"idPais":{"type":"integer","format":"int32"},"cantidad":{"type":"integer","format":"int32"},"tipoFicha":{"type":"string"}}},"AccionReagrupacionPostDTO":{"type":"object","properties":{"idTurno":{"type":"integer","format":"int32"},"idPaisOrigen":{"type":"integer","format":"int32"},"idPaisDestino":{"type":"integer","format":"int32"},"cantidad":{"type":"integer","format":"int32"}}},"AccionModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"idTurnoAtaque":{"type":"integer","format":"int32"},"reagrupacion":{"$ref":"#/components/schemas/ReagrupacionModel"},"combate":{"$ref":"#/components/schemas/CombateModel"}}},"CombateModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"idJugadorAtaque":{"type":"integer","format":"int32"},"idPaisAtk":{"type":"integer","format":"int32"},"fichasAtk":{"type":"integer","format":"int32"},"atkDado1":{"type":"integer","format":"int32"},"atkDado2":{"type":"integer","format":"int32"},"atkDado3":{"type":"integer","format":"int32"},"idJugadorDefensa":{"type":"integer","format":"int32"},"idPaisDef":{"type":"integer","format":"int32"},"fichasDef":{"type":"integer","format":"int32"},"defDado1":{"type":"integer","format":"int32"},"defDado2":{"type":"integer","format":"int32"},"defDado3":{"type":"integer","format":"int32"}}},"ReagrupacionModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"idPaisorigen":{"type":"integer","format":"int32"},"idPaisdestino":{"type":"integer","format":"int32"},"cantidad":{"type":"integer","format":"int32"}}},"AccionCombatePostDTO":{"type":"object","properties":{"idTurno":{"type":"integer","format":"int32"},"idPaisAtk":{"type":"integer","format":"int32"},"idPaisDef":{"type":"integer","format":"int32"}}},"MensajeDto":{"type":"object","properties":{"idJugador":{"type":"integer","format":"int32"},"mensaje":{"type":"string"}}},"UsuarioModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"username":{"type":"string"},"password":{"type":"string"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"CanjeModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"idJugador":{"type":"integer","format":"int32"},"idTurnoRefuerzo":{"type":"integer","format":"int32"},"idtarjeta1":{"type":"integer","format":"int32"},"idtarjeta2":{"type":"integer","format":"int32"},"idtarjeta3":{"type":"integer","format":"int32"},"cantidadFichas":{"type":"integer","format":"int32"}}},"RefuerzoModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"idTurnoRef":{"type":"integer","format":"int32"},"paisJugadorModel":{"$ref":"#/components/schemas/PaisJugadorModel"},"cantidad":{"type":"integer","format":"int32"},"tipoFichas":{"type":"string","enum":["DE_CANJE","CONTINENTE","TARJETA","DE_PAISES"]}}},"RondaModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"idPartida":{"type":"integer","format":"int32"},"numero":{"type":"integer","format":"int32"},"estado":{"type":"string","enum":["EN_PREPARACION","SIN_JUGAR","EN_REFUERZO","EN_ATAQUE","EN_CURSO","FINALIZADO"]},"turnosRefuerzo":{"type":"array","items":{"$ref":"#/components/schemas/TurnoRefuerzoModel"}},"turnosAtaque":{"type":"array","items":{"$ref":"#/components/schemas/TurnoAtaqueModel"}},"jugadores":{"type":"array","items":{"$ref":"#/components/schemas/JugadorModel"}}}},"TurnoAtaqueModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"idJugador":{"type":"integer","format":"int32"},"idTarjeta":{"type":"integer","format":"int32"},"estado":{"type":"string","enum":["EN_PREPARACION","SIN_JUGAR","EN_REFUERZO","EN_ATAQUE","EN_CURSO","FINALIZADO"]},"acciones":{"type":"array","items":{"$ref":"#/components/schemas/AccionModel"}}}},"TurnoRefuerzoModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"idJugador":{"type":"integer","format":"int32"},"estado":{"type":"string","enum":["EN_PREPARACION","SIN_JUGAR","EN_REFUERZO","EN_ATAQUE","EN_CURSO","FINALIZADO"]},"refuerzoModels":{"type":"array","items":{"$ref":"#/components/schemas/RefuerzoModel"}},"canjeModels":{"type":"array","items":{"$ref":"#/components/schemas/CanjeModel"}}}},"FichasDTO":{"type":"object","properties":{"fichasPais":{"type":"integer","format":"int32"},"fichasTarjeta":{"type":"object","additionalProperties":{"type":"integer","format":"int32"}},"fichasContinente":{"type":"object","additionalProperties":{"type":"integer","format":"int32"}}}},"GanadorDTO":{"type":"object","properties":{"idJugador":{"type":"integer","format":"int32"}}},"PaisModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"continente":{"$ref":"#/components/schemas/ContinenteModel"},"idPaisesFronteras":{"type":"array","items":{"type":"integer","format":"int32"}}}},"FronteraJugadorDto":{"type":"object","properties":{"nombreFrontera":{"type":"string"},"fichas":{"type":"integer","format":"int32"},"duenio":{"type":"string"},"idDuenio":{"type":"integer","format":"int32"},"idPais":{"type":"integer","format":"int32"},"color":{"type":"string"}}},"PaisJugadorDto":{"type":"object","properties":{"nombrePaisJugador":{"type":"string"},"fichas":{"type":"integer","format":"int32"},"duenio":{"type":"string"},"idDuenio":{"type":"integer","format":"int32"},"idPais":{"type":"integer","format":"int32"},"fronteras":{"type":"array","items":{"$ref":"#/components/schemas/FronteraJugadorDto"}}}},"MensajeModel":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"idJugador":{"type":"integer","format":"int32"},"mensaje":{"type":"string"},"fechaHora":{"type":"string","format":"date-time"}}}}}}